+++ lib/src/client/channel.dart
@@ -39,6 +39,12 @@ abstract class ClientChannel {
   /// Initiates a new RPC on this connection.
   ClientCall<Q, R> createCall<Q, R>(
       ClientMethod<Q, R> method, Stream<Q> requests, CallOptions options);
+
+  /// Stream of connection state changes
+  ///
+  /// This returns a broadcast stream that can be listened to for connection changes.
+  /// Note: on web channels, this will not yield any values.
+  Stream<ConnectionState> get onConnectionStateChanged;
 }
 
 /// Auxiliary base class implementing much of ClientChannel.
@@ -47,6 +53,8 @@ abstract class ClientChannelBase implements ClientChannel {
   late ClientConnection _connection;
   var _connected = false;
   bool _isShutdown = false;
+  final StreamController<ConnectionState> _connectionStateStreamController =
+      StreamController.broadcast();
 
   ClientChannelBase();
 
@@ -56,6 +64,7 @@ abstract class ClientChannelBase implements ClientChannel {
     _isShutdown = true;
     if (_connected) {
       await _connection.shutdown();
+      await _connectionStateStreamController.close();
     }
   }
 
@@ -64,6 +73,7 @@ abstract class ClientChannelBase implements ClientChannel {
     _isShutdown = true;
     if (_connected) {
       await _connection.terminate();
+      await _connectionStateStreamController.close();
     }
   }
 
@@ -76,6 +86,12 @@ abstract class ClientChannelBase implements ClientChannel {
     if (_isShutdown) throw GrpcError.unavailable('Channel shutting down.');
     if (!_connected) {
       _connection = createConnection();
+      _connection.onStateChanged = (state) {
+        if (_connectionStateStreamController.isClosed) {
+          return;
+        }
+        _connectionStateStreamController.add(state);
+      };
       _connected = true;
     }
     return _connection;
@@ -97,4 +113,8 @@ abstract class ClientChannelBase implements ClientChannel {
     }, onError: call.onConnectionError);
     return call;
   }
+
+  @override
+  Stream<ConnectionState> get onConnectionStateChanged =>
+      _connectionStateStreamController.stream;
 }